Extract Everything3 functions
